iscsi: let conn_close() transfer state of a connection
Current tgtd seems to have a possibility that it has a connection
whose state isn't STATE_CLOSE even after the connection is passed to
conn_close(). Such a connection is confusing because
iscsi_scsi_cmd_done() cannot detect closing of the connection.

This patch lets conn_close() transfer state of a connection to
STATE_CLOSE for avoiding this problem.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>